Split tokenizer into separate file, and add NumericLiteralMatcher
