+++ extension/autocomplete/CMakeLists.txt
@@ -4,7 +4,7 @@ project(AutoCompleteExtension)
 
 include_directories(include)
 
-set(AUTOCOMPLETE_EXTENSION_FILES autocomplete_extension.cpp matcher.cpp)
+set(AUTOCOMPLETE_EXTENSION_FILES autocomplete_extension.cpp matcher.cpp tokenizer.cpp)
 
 build_static_extension(autocomplete ${AUTOCOMPLETE_EXTENSION_FILES})
 set(PARAMETERS "-warnings")
