+++ extension/autocomplete/autocomplete_extension.cpp
@@ -17,6 +17,7 @@
 #include "matcher.hpp"
 #include "duckdb/catalog/default/builtin_types/types.hpp"
 #include "duckdb/main/attached_database.hpp"
+#include "tokenizer.hpp"
 
 namespace duckdb {
 
@@ -252,250 +253,6 @@ static vector<AutoCompleteCandidate> SuggestFileName(ClientContext &context, str
 	});
 	return result;
 }
-
-enum class TokenizeState {
-	STANDARD = 0,
-	IN_COMMENT,
-	QUOTED_IDENTIFIER,
-	STRING_LITERAL,
-	KEYWORD,
-	NUMERIC,
-	OPERATOR
-};
-
-static bool IsSingleByteOperator(char c) {
-	switch(c) {
-	case '(':
-	case ')':
-	case '{':
-	case '}':
-	case '[':
-	case ']':
-	case ',':
-		return true;
-	default:
-		return false;
-	}
-}
-
-static bool CharacterIsInitialNumber(char c) {
-	if (c >= '0' && c <= '9') {
-		return true;
-	}
-	return c == '.';
-}
-
-static bool CharacterIsNumber(char c) {
-	if (CharacterIsInitialNumber(c)) {
-		return true;
-	}
-	switch(c) {
-	case 'e': // exponents
-	case 'E':
-	case '_':
-	case '-':
-		return true;
-	default:
-		return false;
-	}
-}
-
-static bool CharacterIsControlFlow(char c) {
-	switch(c) {
-	case '\'':
-	case '-':
-	case ';':
-	case '"':
-		return true;
-	default:
-		return false;
-	}
-}
-
-static bool CharacterIsKeyword(char c) {
-	if (IsSingleByteOperator(c)) {
-		return false;
-	}
-	if (StringUtil::CharacterIsOperator(c)) {
-		return false;
-	}
-	if (StringUtil::CharacterIsSpace(c)) {
-		return false;
-	}
-	if (CharacterIsControlFlow(c)) {
-		return false;
-	}
-	return true;
-}
-
-static bool CharacterIsOperator(char c) {
-	if (IsSingleByteOperator(c)) {
-		return false;
-	}
-	if (CharacterIsControlFlow(c)) {
-		return false;
-	}
-	return StringUtil::CharacterIsOperator(c);
-}
-
-class BaseTokenizer {
-public:
-	BaseTokenizer(const string &sql, vector<MatcherToken> &tokens) : sql(sql), tokens(tokens) {}
-	virtual ~BaseTokenizer() = default;
-
-	void PushToken(idx_t start, idx_t end) {
-		if (start >= end) {
-			return;
-		}
-		string last_token = sql.substr(start, end - start);
-		tokens.emplace_back(std::move(last_token));
-	}
-
-	bool TokenizeInput() {
-		auto state = TokenizeState::STANDARD;
-
-		idx_t last_pos = 0;
-		idx_t pos_offset = 0;
-		for(idx_t i = 0; i < sql.size(); i++) {
-			auto c = sql[i];
-			switch(state) {
-			case TokenizeState::STANDARD:
-				if (c == '\'') {
-					state = TokenizeState::STRING_LITERAL;
-					last_pos = i;
-					break;
-				}
-				if (c == '"') {
-					state = TokenizeState::QUOTED_IDENTIFIER;
-					last_pos = i;
-					break;
-				}
-				if (c == ';') {
-					// end of statement
-					OnStatementEnd(i);
-					last_pos = i + 1;
-					break;
-				}
-				if (c == '-' && i + 1 < sql.size() && sql[i + 1] == '-') {
-					i++;
-					state = TokenizeState::IN_COMMENT;
-					break;
-				}
-				if (StringUtil::CharacterIsSpace(c)) {
-					// space character - skip
-					last_pos = i + 1;
-					break;
-				}
-				if (IsSingleByteOperator(c)) {
-					// single-byte operator - directly push the token
-					tokens.emplace_back(string(1, c));
-					last_pos = i + 1;
-					break;
-				}
-				if (CharacterIsInitialNumber(c)) {
-					// parse a numeric literal
-					state = TokenizeState::NUMERIC;
-					last_pos = i;
-					break;
-				}
-				if (StringUtil::CharacterIsOperator(c)) {
-					state = TokenizeState::OPERATOR;
-					last_pos = i;
-					break;
-				}
-				state = TokenizeState::KEYWORD;
-				last_pos = i;
-				break;
-			case TokenizeState::NUMERIC:
-				// numeric literal - check if this is still numeric
-				if (!CharacterIsNumber(c)) {
-					// not a number - return to standard state
-					PushToken(last_pos, i);
-					state = TokenizeState::STANDARD;
-					last_pos = i;
-					i--;
-				}
-				break;
-			case TokenizeState::OPERATOR:
-				// operator literal - check if this is still an operator
-				if (!CharacterIsOperator(c)) {
-					// not an operator - return to standard state
-					PushToken(last_pos, i);
-					state = TokenizeState::STANDARD;
-					last_pos = i;
-					i--;
-				}
-				break;
-			case TokenizeState::KEYWORD:
-				// keyword - check if this is still a keyword
-				if (!CharacterIsKeyword(c)) {
-					// not a keyword - return to standard state
-					PushToken(last_pos, i);
-					state = TokenizeState::STANDARD;
-					last_pos = i;
-					i--;
-				}
-				break;
-			case TokenizeState::STRING_LITERAL:
-				if (c == '\'') {
-					if (i + 1 < sql.size() && sql[i + 1] == '\'') {
-						// escaped - skip escape
-						i++;
-					} else {
-						PushToken(last_pos, i + 1);
-						last_pos = i + 1;
-						state = TokenizeState::STANDARD;
-					}
-				}
-				break;
-			case TokenizeState::QUOTED_IDENTIFIER:
-				if (c == '"') {
-					if (i + 1 < sql.size() && sql[i + 1] == '"') {
-						// escaped - skip escape
-						i++;
-					} else {
-						PushToken(last_pos, i + 1);
-						state = TokenizeState::STANDARD;
-					}
-				}
-				break;
-			case TokenizeState::IN_COMMENT:
-				if (c == '\n' || c == '\r') {
-					last_pos = i + 1;
-					state = TokenizeState::STANDARD;
-				}
-				break;
-			default:
-				throw InternalException("unrecognized tokenize state");
-			}
-		}
-
-		// finished processing - check the final state
-		switch(state) {
-		case TokenizeState::STRING_LITERAL:
-			pos_offset = 1;
-			break;
-		case TokenizeState::IN_COMMENT:
-			// no suggestions in comments
-			return false;
-		default:
-			break;
-		}
-		string last_word = sql.substr(last_pos, sql.size() - last_pos);
-		last_pos -= pos_offset;
-		OnLastToken(std::move(last_word), last_pos);
-		return true;
-	}
-
-	virtual void OnStatementEnd(idx_t pos) {
-		tokens.clear();
-	}
-	virtual void OnLastToken(string last_word, idx_t last_pos) = 0;
-
-protected:
-	const string &sql;
-	vector<MatcherToken> &tokens;
-};
 
 class AutoCompleteTokenizer : public BaseTokenizer {
 public:
