+++ extension/autocomplete/grammar/explain.gram
@@ -1,0 +1,1 @@
+ExplainStatement <- 'EXPLAIN'i 'ANALYZE'i? Statement
